Update README.md
# ProjectSetSell-App
Application to Buy products

What is SetSell?

All to often when looking for products to purchase online prices are far to high for used items and sometimes involve meeting in public just to have your time wasted.
SetSell we only post the products the best prices around where all users can expect quality always all pictures are verified and all posts are vetted. 
Please review the attached powerpoint and as far as future add-ons we would like to extend services to allow verified users the ability to sell.